backend/pyproject.toml
[project]
name = "rotini"
version = "0.0.0"
requires-python = ">= 3.10"
dependencies = [
   "uvicorn[standard]",
   "python-multipart",
   "pyjwt",
   "psycopg2",
   "django",
   "djangorestframework",
]

[project.optional-dependencies]
dev = [
    "anyio",
    "black",
    "pylint",
    "pylint_django",
    "pytest-django",
    "pytest",
]

[tool.setuptools]
packages = ["rotini"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE="base.settings"
pythonpath=[
    ".",
    "./rotini",
]

[tool.pylint.'MASTER']
load-plugins="pylint_django"

